+++ lib/src/actors/spv_actor.dart
@@ -12,7 +12,7 @@ import '../storage/wallet_storage.dart';
 import '../utils/beef.dart';
 import '../utils/bump.dart';
 import '../models/invoice_output_spec.dart';
-import 'spv_messages.dart' hide ValidateBEEFMessage, BEEFValidationResult;
+import 'spv_messages.dart' hide ValidateBEEFMessage;
 import 'wallet_messages.dart';
 // Both message libraries declare RetrieveMerkleProofMessage; ARCActor handles
 // the one in wallet_messages.dart (bead libspiffy-0lx).